Membetulkan error

+++ database/migrations/2026_06_20_144047_create_inventories_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('stock');
 
             $table->integer('isipc');
-            $table->integer('hrgktn');
+            $table->bigInteger('hrgktn');
 
             $table->timestamps();
         });
